Move test to its own file and make helper static
